requirements-dev.txt
pytest==7.4.3
pytest-mock==3.12.0
moto==4.2.14
black==23.12.1
ruff==0.1.8
mypy==1.8.0
mutmut==2.4.3
boto3-stubs[dynamodb]==1.34.0